# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CheapDataHub VTU & Bill Payment API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services/data:
    post:
      operationId: buyData
      tags: [services]
      summary: Buy data plan for a phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataPurchaseInput"
      responses:
        "200":
          description: Data purchase result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
  /services/airtime:
    post:
      operationId: buyAirtime
      tags: [services]
      summary: Buy airtime for a phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AirtimePurchaseInput"
      responses:
        "200":
          description: Airtime purchase result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
  /services/cable:
    post:
      operationId: buyCable
      tags: [services]
      summary: Subscribe to cable TV plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CablePurchaseInput"
      responses:
        "200":
          description: Cable purchase result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
  /services/electricity:
    post:
      operationId: buyElectricity
      tags: [services]
      summary: Purchase electricity tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ElectricityPurchaseInput"
      responses:
        "200":
          description: Electricity purchase result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"

  # ─── Wallet ─────────────────────────────────────────────────────────────────
  /wallet/fund/initialize:
    post:
      operationId: initializeWalletFunding
      tags: [wallet]
      summary: Initialize a wallet funding request and get payment details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletFundInput"
      responses:
        "200":
          description: Payment initialization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletFundResponse"
...
  /webhooks/paystack:
    post:
      operationId: paystackWebhook
      tags: [webhooks]
      summary: Paystack charge.success webhook handler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookPayload"
      responses:
        "200":
          description: Webhook acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Admin ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
...
    DataPurchaseInput:
      type: object
      required: [phone, plan_id, pin]
      properties:
        phone:
          type: string
        plan_id:
          type: string
        network:
          type: string
        pin:
          type: string
    AirtimePurchaseInput:
      type: object
      required: [phone, network, amount, pin]
      properties:
        phone:
          type: string
        network:
          type: string
        amount:
          type: number
        pin:
          type: string
    CablePurchaseInput:
      type: object
      required: [smart_card_number, cable_provider, plan_id, pin]
      properties:
        smart_card_number:
          type: string
        cable_provider:
          type: string
        plan_id:
          type: string
        pin:
          type: string
    ElectricityPurchaseInput:
      type: object
      required: [meter_number, disco, amount, meter_type, pin]
      properties:
        meter_number:
          type: string
        disco:
          type: string
        amount:
          type: number
        meter_type:
          type: string
        pin:
          type: string
    ServiceResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        token:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
        new_balance:
          type: ["number", "null"]
    WalletFundInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
    WalletFundResponse:
      type: object
      required: [base_amount, processing_fee, total_amount, reference]
      properties:
        base_amount:
          type: number
        processing_fee:
          type: number
        total_amount:
          type: number
        reference:
          type: string
        payment_instruction:
          type: string
...
    WebhookPayload:
      type: object
      additionalProperties: true
      properties:
        event:
          type: string
        data:
          type: object
          additionalProperties: true