# lib/api-zod/src/generated/types/airtimePurchaseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CheapDataHub VTU & Bill Payment API
 * OpenAPI spec version: 0.1.0
 */

export interface AirtimePurchaseInput {
  phone: string;
  network: string;
  amount: number;
  pin: string;
}