# lib/api-zod/src/generated/types/cablePurchaseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CheapDataHub VTU & Bill Payment API
 * OpenAPI spec version: 0.1.0
 */

export interface CablePurchaseInput {
  smart_card_number: string;
  cable_provider: string;
  plan_id: string;
  pin: string;
}