# lib/api-zod/src/generated/types/dataPurchaseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CheapDataHub VTU & Bill Payment API
 * OpenAPI spec version: 0.1.0
 */

export interface DataPurchaseInput {
  phone: string;
  plan_id: string;
  network?: string;
  pin: string;
}