# lib/api-zod/src/generated/types/serviceResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CheapDataHub VTU & Bill Payment API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceResponse {
  success: boolean;
  message: string;
  /** @nullable */
  token?: string | null;
  /** @nullable */
  reference?: string | null;
  /** @nullable */
  new_balance?: number | null;
}